GH-3804: Migrate SMB unit tests to Testcontainers
Migrate to another docker image, elswork/samba

The previous docker image did not have proper versioned tags. We had to
use `latest` tag and this is not a very robust approach, so to avoid it
we have to either use a particular sha digest (which has drawbacks as
puts a lock on a particular CPU architecture) or use some different
image which is a much better solution.